+++ src/countries/om.ts
@@ -1,0 +1,21 @@
+import { Country } from './Country'
+
+export const om: Country = {
+	code: 'om',
+	name: {
+		en: 'Oman',
+		ar: 'عُمان',
+		de: 'Oman',
+		sv: 'Oman',
+		da: 'Oman',
+		sq: 'Omani',
+		pt: 'Omã',
+		tr: 'Umman',
+		fa: 'عمان',
+		uk: 'Оман',
+		xa: 'السلام السلطاني',
+		xt: 'السلام السلطاني',
+	},
+	flag: '🇴🇲',
+	beta: true,
+}