+++ src/countries/th.ts
@@ -1,0 +1,21 @@
+import { Country } from './Country'
+
+export const th: Country = {
+	code: 'th',
+	name: {
+		en: 'Thailand',
+		ar: 'تايلاند',
+		de: 'Thailand',
+		sv: 'Thailand',
+		da: 'Thailand',
+		sq: 'Tajlanda',
+		pt: 'Tailândia',
+		tr: 'Tayland',
+		fa: 'تایلند',
+		uk: 'Таїланд',
+		xa: 'เพลงชาติไทย',
+		xt: 'เพลงชาติไทย',
+	},
+	flag: '🇹🇭',
+	beta: true,
+}