+++ src/countries/es.ts
@@ -1,0 +1,21 @@
+import { Country } from './Country'
+
+export const es: Country = {
+	code: 'es',
+	name: {
+		en: 'Spain',
+		ar: 'إسبانيا',
+		de: 'Spanien',
+		sv: 'Spanien',
+		da: 'Spanien',
+		sq: 'Spanja',
+		pt: 'Espanha',
+		tr: 'İspanya',
+		fa: 'اسپانیا',
+		uk: 'Іспанія',
+		xa: 'La Marcha Real',
+		xt: 'La Marcha Real',
+	},
+	flag: '🇪🇸',
+	beta: true,
+}