+++ src/countries/va.ts
@@ -1,0 +1,21 @@
+import { Country } from './Country'
+
+export const va: Country = {
+	code: 'va',
+	name: {
+		en: 'Vatican City',
+		ar: 'الفاتيكان',
+		de: 'Vatikanstadt',
+		sv: 'Vatikanstaten',
+		da: 'Vatikanstaten',
+		sq: 'Vatikani',
+		pt: 'Vaticano',
+		tr: 'Vatikan',
+		fa: 'واتیکان',
+		uk: 'Ватикан',
+		xa: 'Inno e Marcia Pontificale',
+		xt: 'Inno e Marcia Pontificale',
+	},
+	flag: '🇻🇦',
+	beta: true,
+}