+++ src/countries/at.ts
@@ -1,0 +1,20 @@
+import { Country } from './Country'
+
+export const at: Country = {
+	code: 'at',
+	name: {
+		en: 'Austria',
+		ar: 'النمسا',
+		de: 'Österreich',
+		sv: 'Österrike',
+		da: 'Østrig',
+		sq: 'Austria',
+		pt: 'Áustria',
+		tr: 'Avusturya',
+		fa: 'اتریش',
+		uk: 'Австрія',
+		xa: 'Land der Berge, Land am Strome',
+		xt: 'Land der Berge, Land am Strome',
+	},
+	flag: '🇦🇹',
+}