+++ src/countries/be.ts
@@ -1,0 +1,20 @@
+import { Country } from './Country'
+
+export const be: Country = {
+	code: 'be',
+	name: {
+		en: 'Belgium',
+		ar: 'بلجيكا',
+		de: 'Belgien',
+		sv: 'Belgien',
+		da: 'Belgien',
+		sq: 'Belgjika',
+		pt: 'Bélgica',
+		tr: 'Belçika',
+		fa: 'بلژیک',
+		uk: 'Бельгія',
+		xa: 'La Brabançonne',
+		xt: 'La Brabançonne',
+	},
+	flag: '🇧🇪',
+}