+++ src/countries/ch.ts
@@ -1,0 +1,20 @@
+import { Country } from './Country'
+
+export const ch: Country = {
+	code: 'ch',
+	name: {
+		en: 'Switzerland',
+		ar: 'سويسرا',
+		de: 'Schweiz',
+		sv: 'Schweiz',
+		da: 'Schweiz',
+		sq: 'Zvicra',
+		pt: 'Suíça',
+		tr: 'İsviçre',
+		fa: 'سوئیس',
+		uk: 'Швейцарія',
+		xa: 'Schweizerpsalm',
+		xt: 'Schweizerpsalm',
+	},
+	flag: '🇨🇭',
+}