+++ src/countries/gr.ts
@@ -1,0 +1,20 @@
+import { Country } from './Country'
+
+export const gr: Country = {
+	code: 'gr',
+	name: {
+		en: 'Greece',
+		ar: 'اليونان',
+		de: 'Griechenland',
+		sv: 'Grekland',
+		da: 'Grækenland',
+		sq: 'Greqia',
+		pt: 'Grécia',
+		tr: 'Yunanistan',
+		fa: 'یونان',
+		uk: 'Греція',
+		xa: 'Ύμνος εις την Ελευθερίαν',
+		xt: 'Ύμνος εις την Ελευθερίαν',
+	},
+	flag: '🇬🇷',
+}