+++ src/countries/it.ts
@@ -1,0 +1,20 @@
+import { Country } from './Country'
+
+export const it: Country = {
+	code: 'it',
+	name: {
+		en: 'Italy',
+		ar: 'إيطاليا',
+		de: 'Italien',
+		sv: 'Italien',
+		da: 'Italien',
+		sq: 'Italia',
+		pt: 'Itália',
+		tr: 'İtalya',
+		fa: 'ایتالیا',
+		uk: 'Італія',
+		xa: 'Il Canto degli Italiani',
+		xt: 'Il Canto degli Italiani',
+	},
+	flag: '🇮🇹',
+}