+++ src/countries/lu.ts
@@ -1,0 +1,20 @@
+import { Country } from './Country'
+
+export const lu: Country = {
+	code: 'lu',
+	name: {
+		en: 'Luxembourg',
+		ar: 'لوكسمبورغ',
+		de: 'Luxemburg',
+		sv: 'Luxemburg',
+		da: 'Luxembourg',
+		sq: 'Luksemburgu',
+		pt: 'Luxemburgo',
+		tr: 'Lüksemburg',
+		fa: 'لوکزامبورگ',
+		uk: 'Люксембург',
+		xa: 'Ons Heemecht',
+		xt: 'Ons Heemecht',
+	},
+	flag: '🇱🇺',
+}