+++ src/countries/nl.ts
@@ -1,0 +1,20 @@
+import { Country } from './Country'
+
+export const nl: Country = {
+	code: 'nl',
+	name: {
+		en: 'Netherlands',
+		ar: 'هولندا',
+		de: 'Niederlande',
+		sv: 'Nederländerna',
+		da: 'Nederlandene',
+		sq: 'Holanda',
+		pt: 'Países Baixos',
+		tr: 'Hollanda',
+		fa: 'هلند',
+		uk: 'Нідерланди',
+		xa: 'Het Wilhelmus',
+		xt: 'Het Wilhelmus',
+	},
+	flag: '🇳🇱',
+}