+++ src/countries/pl.ts
@@ -1,0 +1,20 @@
+import { Country } from './Country'
+
+export const pl: Country = {
+	code: 'pl',
+	name: {
+		en: 'Poland',
+		ar: 'بولندا',
+		de: 'Polen',
+		sv: 'Polen',
+		da: 'Polen',
+		sq: 'Polonia',
+		pt: 'Polónia',
+		tr: 'Polonya',
+		fa: 'لهستان',
+		uk: 'Польща',
+		xa: 'Mazurek Dąbrowskiego',
+		xt: 'Mazurek Dąbrowskiego',
+	},
+	flag: '🇵🇱',
+}